move tooltip to new withCoordsv2
