+++ src/universal/utils/isKeyboardEvent.js
@@ -1,0 +1,3 @@
+const isKeyboardEvent = (e) => e && e.key;
+
+export default isKeyboardEvent;
